requirements.txt
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.1
click==7.1.2
cryptography==3.0
decorator==4.4.2
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.20
pylint==2.5.3
python-telegram-bot==12.8
redis==3.5.3
redis-server==5.0.7
six==1.15.0
toml==0.10.1
tornado==6.0.4
Werkzeug==1.0.1
wrapt==1.12.1
